requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.9.0
google-generativeai==0.7.2
razorpay==1.4.1
python-multipart==0.0.9
bcrypt==4.0.1
reportlab==4.1.0
